initial oracle commit
This commit contains an initial version of oracle. It contains the bare
minimum for something to be demo'd.

It only works if:
* the schema and table names match case.
* the column names are lowercase in PG.
* The only tested types are integers and strings.

No escaping work has yet been done - but that will come - this makes the
PR a tad easier to read.

Unfortunately, CI w/ Oracle is not cooporating with me at the moment -
I'll tackle that later.