LMS: Add Helm Chart (#20)
* LMS: Add Helm Chart

This commit adds the production helm chart for the LMS. It will
deploy the LMS and Orchestrator by itself with full customization
over how you want to configure it. There is an optional service
monitor that can be installed if running with the kubernetes
operator.

* LMS: Remove release as label from chart

This commit removes the value release from the chart as something
that can be set. It wasn't really needed and users and supplement
with any label of their choosing.