fetch: fixed direct-copy functionality and disallow compression for COPY modes (#52)
Previously, compression was applied for a default for both the non-live mode and direct copy without the live flag. This made the writer for copy compressed by default, leading to an issue when copying. This change makes it so that we don't allow compression while copying and updating the default logic so that we can properly pass in compression opts.

Release Note: None