cmdutil: refined logging scheme by adding timestamps, structured logging as a default, and option to write to a log file (#91)
Resolves: CC-26366, CC-26376
Release Note: Logging for MOLT tools (Fetch and Verify) now defaults to structured JSON logging, which contains at least the message, level, and formatted time string. Also added a `--use-console-writer` flag that can be enabled to use prettier console writing that has colored and human-readable log output. Structured logging is preferred as a default because it is easier to parse with log ingestion tools and is more efficient for the application to output. You can also set `--log-file` to write your task/data logs to specified log files. If the flag is left empty, the MOLT tool will only write to stdout, which is more efficient.